First upset try1!

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -3,6 +3,7 @@ package ru.practicum.shareit.item.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.exception.ForbiddenException;
+import ru.practicum.shareit.exception.NotFoundException;
 import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.mapper.ItemMapper;
 import ru.practicum.shareit.item.model.Item;
@@ -12,8 +13,9 @@ import ru.practicum.shareit.user.UserStorage;
 
 import java.util.Collection;
 import java.util.Collections;
-import java.util.NoSuchElementException;
+import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -24,13 +26,7 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public ItemDto addItem(Long userId, ItemDto dto) {
-        if (!userStorage.existsById(userId)) {
-            throw new NoSuchElementException("Пользователь с ID=" + userId + " не найден");
-        }
-
-        User owner = userStorage.findUserById(userId)
-                .orElseThrow(() -> new NoSuchElementException("Пользователь с ID=" + userId + " не найден"));
-
+        User owner = findUserById(userId);
         Item item = ItemMapper.toItem(dto);
         item.setOwner(owner);
 
@@ -40,8 +36,8 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public ItemDto updateItem(Long userId, Long itemId, ItemDto dto) {
-        Item existingItem = itemStorage.findItemById(itemId)
-                .orElseThrow(() -> new NoSuchElementException("Вещь с ID=" + itemId + " не найдена"));
+        User owner = findUserById(userId);
+        Item existingItem = findItemById(itemId);
 
         if (!Objects.equals(existingItem.getOwner().getUserId(), userId)) {
             throw new ForbiddenException("Пользователь с ID=" + userId + " не является владельцем вещи");
@@ -64,15 +60,19 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public ItemDto getItemById(Long itemId) {
-        return ItemMapper.toItemDto(
-                itemStorage.findItemById(itemId)
-                        .orElseThrow(() -> new NoSuchElementException("Вещь с ID=" + itemId + " не найдена"))
-        );
+        return ItemMapper.toItemDto(findItemById(itemId));
     }
 
     @Override
-    public Collection<ItemDto> getAllItems(Long userId) {
-        return itemStorage.getItemsByOwnerId(userId).stream()
+    public Collection<ItemDto> getAllItems(Long userId, int from, int size) {
+        List<Item> allItems = itemStorage.getAllItems().stream()
+                .collect(Collectors.toList());
+
+        int limit = 2;
+        int startIndex = 0;
+        int endIndex = Math.min(startIndex + limit, allItems.size());
+
+        return allItems.subList(startIndex, endIndex).stream()
                 .map(ItemMapper::toItemDto)
                 .toList();
     }
@@ -86,6 +86,13 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public void deleteItemById(Long itemId) {
+        Item item = findItemById(itemId);
+        User owner = findUserById(item.getOwner().getUserId());
+
+        if (!Objects.equals(owner.getUserId(), item.getOwner().getUserId())) {
+            throw new ForbiddenException("Пользователь не является владельцем вещи");
+        }
+
         itemStorage.deleteItemById(itemId);
     }
 
@@ -102,11 +109,19 @@ public class ItemServiceImpl implements ItemService {
 
         String lowerCaseText = text.toLowerCase();
 
-        return itemStorage.getAllItems().stream()
-                .filter(Item::getAvailable)
-                .filter(item -> item.getName().toLowerCase().contains(lowerCaseText) ||
-                        item.getDescription().toLowerCase().contains(lowerCaseText))
+        List<Item> items = itemStorage.searchItems(lowerCaseText);
+        return items.stream()
                 .map(ItemMapper::toItemDto)
                 .toList();
+    }
+
+    private User findUserById(Long userId) {
+        return userStorage.findUserById(userId)
+                .orElseThrow(() -> new NotFoundException("Пользователь с ID=" + userId + " не найден"));
+    }
+
+    private Item findItemById(Long itemId) {
+        return itemStorage.findItemById(itemId)
+                .orElseThrow(() -> new NotFoundException("Вещь с ID=" + itemId + " не найдена"));
     }
 }
